requirements.txt
# Core Dashboard Framework
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
plotly==5.18.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# Excel File Handling
openpyxl==3.1.2
xlrd==2.0.1

# Date/Time Processing
python-dateutil==2.8.2

# Caching
Flask-Caching==2.1.0

# Server
gunicorn==21.2.0
Flask==3.0.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Statistical Analysis
statsmodels==0.14.1

# Geographic Data Processing (if needed)
geopandas==0.14.1
folium==0.15.1

# Development Tools
black==23.11.0
flake8==6.1.0
pytest==7.4.3
jupyter==1.0.0